requirements.txt
pandas>=2.0
python-dotenv==1.0.0
pdfplumber==0.10.2
openpyxl==3.1.2
xlrd==2.0.1